Fixes change-password when condition (bsc#1085357)
Fixes when condition which generates an UndefinedError if the previous
task is skipped. When skipped, the registered variable
replication_result does not contain rc, the double-braced expression of
the when condition is evaluated first (in order of precedence), and an
UndefinedError is generated. By removing the double braces, the
order of precedence is restored and replication_result.rc is
evaluated last.

Change-Id: Id1ef0e6a0bbbf18af9e0741d4ada0c66fa6caedc