Get ca bundle location from central variable (bsc#1084151)
https://gerrit.suse.provo.cloud//#/c/3689/ introduces a new variable for
the path to the bundle of trusted ca certificates. This change makes
use of that variable to avoid to adjust the file name across many places in the
future.

This commit also removes the unused post_java_install_restore task.
The task is currently not executed by anything. And with the above
referenced change it will also be unneeded in the future.

Depends-On: Ia9a1113e7b59664c0acfc5aeb14dd0e5309b3c76
Change-Id: I671a40701376de7ab187c3070f7de0bbc14e1926